With c-ares, we can get to its ares_getnameinfo lookup conclusion point via one of two paths. First, on a cache or local file lookup or on an error path, we can arrive here without going through a socket state and subsequent data event, which means we are still in the context of a 'CFHostStartInfoResolution' down call. Second, on a DNS server lookup, we can arrive here via a socket state and subsequent data event, which means we are getting scheduled and down called via the run loop dispatch.
In the first case, the host object is locked and any call into _GetNameinfoCallBackWithFree would deadlock, based on the Mach legacy code base.
In the second case, the host object is unlocked and any such call will not deadlock.

Consequently, we need to handle these two cases distinctly and with care. The distinguishing factor will be that the first case uses the
special null lookup and should NOT lock; whereas, the second case does not and SHOULD lock.

The _GetNameinfoCallBackWithFree has been reworked with an expanded API and renamed _GetNameInfoCallBackWithFreeAndWithShouldLock to take not
only an optional resource deallocation callback but also a Boolean to indicate whether the host object should be locked within the function.

The other nuance we need to deal with in the first case is that normally host->_lookup will not get assigned until the call to
_CreateNameLookup unwinds in _CreateLookup_NoLock. However, the down call into _GetNameInfoCallBackWithFreeAndWithShouldLock (from the Mach
legacy code base) assumes host->_lookup has already been assigned. We will have to manually assign it in that case.