Note that in certain cases, at least with TCP sockets, there may be a small single-digit to mid double-digit millisecond window of time between when the socket is connecting and the select/poll signal that the socket is ready for reading and when the first data shows up on the socket for actual reading. In that intervening time, read attempts on a non-blocking socket will return EAGAIN.
At least on Linux, experience has shown that the EAGAIN-memset-continue and subsequent _kCFStreamSocketReadPrivateMode approach found in _SocketStreamRead does not actually work to
resolve this situation. With that approach, the socket either experiences a time out (which never happens on iOS/macOS/tvOS) or the process incorporating the socket stream deadlocks
indefinitely for a read/recv retry that never arrives (which likewise never happens on iOS/macOS/tvOS). Consequently, given the short windows involved and the seeming-ineffectiveness of
the current solution for this, we employ a time-bound retry, continuing _CFSocketRecv, until the receive is successful or until the retry timeout expires.